Fix wrong pairing between files in SIFID
We encountered a bug when different images were compared to each other. The root cause is that `glob(...)` doesn't always keep order, so `files1` and `files2` were ordered differently.
A fix for that bug was sorting `files1` and `files2`. We assume that in each list, all the paths are in the same directory, and that files have the same name between the two lists - so sorting the two lists would guarantee a matching order.